Namespace and multiprocessing
- Clean-up namespace by setting dir to __all__
- Move multiprocessing stuff to new private module _multiprocessing.
